+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.55.0] - 2024-10-21
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): Activation page at `/users/activate/[activationCode]`.
+- [Marek Kurańda](https://github.com/mjkuranda): Confirmation modal for all objects at manage page.
+- [Marek Kurańda](https://github.com/mjkuranda): Object count at every object manage page.
+
 ## [1.54.3] - 2024-10-21
 ### Fixed
 - [Marek Kurańda](https://github.com/mjkuranda): Loading dish images.